# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Post Office Attendance Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [present, absent, late, half_day, holiday, leave]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceListResponse"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create/update attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRequest"
      responses:
        "201":
          description: Attendance saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}:
    put:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRequest"
      responses:
        "200":
          description: Updated attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/today:
    get:
      operationId: getTodayAttendance
      tags: [attendance]
      summary: Get today's attendance summary
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Today attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayAttendanceSummary"
  /attendance/monthly-sheet:
    get:
      operationId: getMonthlySheet
      tags: [attendance]
      summary: Get monthly attendance sheet
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly attendance sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlySheetResponse"
  /attendance/punch:
    post:
      operationId: punchAttendance
      tags: [attendance]
      summary: Manual punch in/out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PunchRequest"
      responses:
        "200":
          description: Punch recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /reports/attendance:
    get:
      operationId: getAttendanceReport
      tags: [reports]
      summary: Attendance report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceReport"
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeCode:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, half_day, holiday, leave]
        inTime1:
          type: string
          nullable: true
        outTime1:
          type: string
          nullable: true
        workHours1:
          type: number
          nullable: true
        inTime2:
          type: string
          nullable: true
        outTime2:
          type: string
          nullable: true
        workHours2:
          type: number
          nullable: true
        totalHours:
          type: number
          nullable: true
        overtimeHours:
          type: number
          nullable: true
        shiftName:
          type: string
          nullable: true
        source:
          type: string
          enum: [biometric, manual, system]
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, employeeId, employeeName, employeeCode, branchId, branchName, date, status, source, createdAt]
    AttendanceListResponse:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [records, total, page, limit]
    CreateAttendanceRequest:
      type: object
      properties:
        employeeId:
          type: integer
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, half_day, holiday, leave]
        inTime1:
          type: string
          nullable: true
        outTime1:
          type: string
          nullable: true
        inTime2:
          type: string
          nullable: true
        outTime2:
          type: string
          nullable: true
        overtimeHours:
          type: number
          nullable: true
        source:
          type: string
          enum: [biometric, manual, system]
        remarks:
          type: string
          nullable: true
      required: [employeeId, date, status, source]
    PunchRequest:
      type: object
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [in, out]
        time:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
      required: [employeeId, type]
    TodayAttendanceSummary:
      type: object
      properties:
        date:
          type: string
        totalEmployees:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        halfDay:
          type: integer
        onLeave:
          type: integer
        notMarked:
          type: integer
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
      required: [date, totalEmployees, present, absent, late, halfDay, onLeave, notMarked, records]
    MonthlySheetRow:
      type: object
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeCode:
          type: string
        designation:
          type: string
        dailyStatus:
          type: array
          items:
            type: object
            properties:
              day:
                type: integer
              status:
                type: string
              inTime:
                type: string
                nullable: true
              outTime:
                type: string
                nullable: true
              hours:
                type: number
                nullable: true
            required: [day, status]
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        halfDays:
          type: integer
        leaveDays:
          type: integer
        holidayDays:
          type: integer
        totalWorkHours:
          type: number
        overtimeHours:
          type: number
      required: [employeeId, employeeName, employeeCode, designation, dailyStatus, presentDays, absentDays, lateDays, halfDays, leaveDays, holidayDays, totalWorkHours, overtimeHours]
    MonthlySheetResponse:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        daysInMonth:
          type: integer
        branchName:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/MonthlySheetRow"
      required: [month, year, daysInMonth, branchName, rows]
    AttendanceReport:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        totalRecords:
          type: integer
        summary:
          type: object
          properties:
            present:
              type: integer
            absent:
              type: integer
            late:
              type: integer
            halfDay:
              type: integer
            leave:
              type: integer
            holiday:
              type: integer
          required: [present, absent, late, halfDay, leave, holiday]
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
      required: [startDate, endDate, totalRecords, summary, records]
...