# lib/api-zod/src/generated/types/monthlySheetRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Post Office Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlySheetRowDailyStatusItem } from "./monthlySheetRowDailyStatusItem";

export interface MonthlySheetRow {
  employeeId: number;
  employeeName: string;
  employeeCode: string;
  designation: string;
  dailyStatus: MonthlySheetRowDailyStatusItem[];
  presentDays: number;
  absentDays: number;
  lateDays: number;
  halfDays: number;
  leaveDays: number;
  holidayDays: number;
  totalWorkHours: number;
  overtimeHours: number;
}